+++ src/packages/piral-cli/src/apps/common/scaffoldPiral.ts
@@ -1,7 +1,7 @@
+import { join, dirname, relative } from 'path';
 import { PiletLanguage, getLanguageExtension, getDevDependencies } from './language';
 import { getPiletsInfo } from './package';
 import { ForceOverwrite, createFileIfNotExists } from './io';
-import { join, dirname, relative } from 'path';
 
 function getPiralAppContent(language: PiletLanguage) {
   const extension = getLanguageExtension(language);
@@ -20,12 +20,62 @@ function getPiralAppContent(language: PiletLanguage) {
 </html>`;
 }
 
+function getPiralMockContent() {
+  return `const request = require('request');
+
+function isPiletQuery(content) {
+  try {
+    const q = JSON.parse(content).query.replace(/\\s+/g, ' ');
+    return q.indexOf('query initialData { pilets {') !== -1;
+  } catch (e) {
+    return false;
+  }
+}
+
+// Place a script here to "redirect" a standard API to some GraphQL.
+const apiService = '';// 'https://feed.piral.io/api/v1/pilet/sample';
+
+module.exports = function(_, req, res) {
+  if ((req.path = '/' && req.method === 'POST' && isPiletQuery(req.content))) {
+    if (apiService) {
+      return new Promise(resolve => {
+        request.get(apiService, (_1, _2, body) => {
+          const response = res({
+            content: JSON.stringify({
+              data: {
+                pilets: JSON.parse(body).items,
+              },
+            }),
+          });
+          resolve(response);
+        });
+      });
+    } else {
+      return res({
+        content: JSON.stringify({
+          data: {
+            pilets: [],
+          },
+        }),
+      });
+    }
+  }
+};
+`;
+}
+
 function getPiralRootModuleContent(packageName: string) {
   switch (packageName) {
     case 'piral':
-      return `import { renderInstance, buildLayout } from 'piral';
+      return `import * as React from 'react';
+import { renderInstance, buildLayout } from 'piral';
 
-renderInstance({ layout: buildLayout() });
+renderInstance({
+  layout: buildLayout()
+    .withError(({ type }) => (
+      <span style={{ color: 'red', fontWeight: 'bold' }}>Error: {type}</span>
+    )),
+});
 `;
     case 'piral-core':
     default:
@@ -49,6 +99,18 @@ export function getPiralPackage(app: string, language: PiletLanguage) {
       return {
         ...baseData,
         typings: 'lib/index.d.ts',
+        scripts: {
+          build: 'npm run build:deploy && npm run build:pilets',
+          'build:deploy': 'piral build',
+          'build:pilets': 'tsc',
+        },
+      };
+    case PiletLanguage.js:
+      return {
+        ...baseData,
+        scripts: {
+          build: 'piral build',
+        },
       };
   }
 
@@ -63,7 +125,9 @@ export async function scaffoldPiralSourceFiles(
   forceOverwrite: ForceOverwrite,
 ) {
   const src = dirname(join(root, app));
+  const mocks = join(src, 'mocks');
   await createFileIfNotExists(root, app, getPiralAppContent(language), forceOverwrite);
+  await createFileIfNotExists(mocks, 'backend.js', getPiralMockContent(), forceOverwrite);
 
   switch (language) {
     case PiletLanguage.ts:
@@ -97,7 +161,7 @@ export async function scaffoldPiralSourceFiles(
         src,
         'index.tsx',
         `${getPiralRootModuleContent(packageName)}
-export * from '${packageName}';
+export * from '${packageName}/lib/types';
 `,
         forceOverwrite,
       );