+++ src/packages/piral-cli/src/apps/new-pilet.test.ts
@@ -12,7 +12,7 @@ describe('New Pilet Command', () => {
   it('scaffolding in an empty directory works', async () => {
     jest.setTimeout(60000);
     const dir = createTempDir();
-    await newPilet(dir);
+    await newPilet(dir, { skipInstall: true });
     expect(existsSync(resolve(dir, 'node_modules/piral/package.json'))).toBeTruthy();
     expect(existsSync(resolve(dir, 'package.json'))).toBeTruthy();
     expect(existsSync(resolve(dir, 'tsconfig.json'))).toBeTruthy();
@@ -25,6 +25,7 @@ describe('New Pilet Command', () => {
     const dir = createTempDir();
     await newPilet(dir, {
       language: PiletLanguage.js,
+      skipInstall: true,
     });
     expect(existsSync(resolve(dir, 'node_modules/piral/package.json'))).toBeTruthy();
     expect(existsSync(resolve(dir, 'package.json'))).toBeTruthy();