+++ src/packages/piral-core/src/helpers.tsx
@@ -1,6 +1,6 @@
 import { AvailableDependencies } from 'react-arbiter';
 import { createBrowserHistory } from 'history';
-import { globalDependencies } from './modules';
+import { globalDependencies, getLocalDependencies } from './modules';
 import { defaultBreakpoints, getUserLocale, getCurrentLayout, defaultLayouts } from './utils';
 import { DefaultDashboard, DefaultLoader, DefaultErrorInfo } from './components/default';
 import {
@@ -18,12 +18,24 @@ import {
  * Creates a dependency getter that sets the shared dependencies explicitly.
  * Overrides the potentially set shared dependencies from the Piral CLI, but
  * keeps all global dependencies such as react, react-dom, ...
- * @param dependencies The shared dependencies to declare.
+ * @param sharedDependencies The shared dependencies to declare.
  */
 export function setSharedDependencies(sharedDependencies: AvailableDependencies) {
   const dependencies = {
     ...globalDependencies,
     ...sharedDependencies,
+  };
+  return () => dependencies;
+}
+
+/**
+ * Creates a dependency getter that extends the shared dependencies with additional dependencies.
+ * @param additionalDependencies The additional dependencies to declare.
+ */
+export function extendSharedDependencies(additionalDependencies: AvailableDependencies) {
+  const dependencies = {
+    ...getLocalDependencies(),
+    ...additionalDependencies,
   };
   return () => dependencies;
 }