+++ src/pages/docs/src/scripts/documentation/Content.tsx
@@ -1,15 +1,10 @@
 import * as React from 'react';
 import { Section, ResponsiveContent, Md, EditSection, Callout } from '../components';
 
-export interface ContentProps {
-  ref?: React.Ref<HTMLDivElement>;
-}
+export interface ContentProps {}
 
-export const Content: React.FC<ContentProps> = React.forwardRef((_, ref) => (
+export const Content = React.forwardRef<HTMLDivElement, ContentProps>((_, ref) => (
   <ResponsiveContent ref={ref}>
-    <Callout title="Pilet Guideline" type="warning" icon="exclamation-triangle" to="/guidelines/static-piral-instance">
-      <p>Looking for help on creating a Piral instance? Check out our guideline for an easy approach.</p>
-    </Callout>
     {/* start:auto-generated */}
     <Section id="section-introduction" title="Introduction">
       <Md>{require('../../../../../../docs/introduction.md')}</Md>
@@ -31,17 +26,10 @@ export const Content: React.FC<ContentProps> = React.forwardRef((_, ref) => (
       <Md>{require('../../../../../../docs/browsers.md')}</Md>
       <EditSection link="browsers.md" />
     </Section>
-    <Section id="section-development" title="Development">
-      <Md>{require('../../../../../../docs/development.md')}</Md>
-      <EditSection link="development.md" />
-    </Section>
-    <Section id="section-history" title="History">
+    <Section id="section-project-history" title="Project History">
       <Md>{require('../../../../../../docs/history.md')}</Md>
       <EditSection link="history.md" />
     </Section>
     {/* end:auto-generated */}
-    <Callout title="Pilet Guideline" type="warning" icon="exclamation-triangle" to="/guidelines/first-pilet">
-      <p>Looking for help on developing pilets? Check out our guidelines for developing pilets.</p>
-    </Callout>
   </ResponsiveContent>
 ));